requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch==1.13.1
torchvision==0.14.1
tqdm>=4.41.0
protobuf<4.21.3

# plotting ------------------------------------

pandas>=1.1.4
seaborn>=0.11.0

# deep_sort -----------------------------------

easydict

# torchreid -----------------------------------

Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio

# Export --------------------------------------
#onnx>=1.9.0  # ONNX export
#onnx-simplifier>=0.4.1  # ONNX simplifier
#tensorflow>=2.4.1  # TFLite export (or tensorflow-cpu, tensorflow-aarch64)
#openvino-dev  # OpenVINO export
#git+https://github.com/PINTO0309/openvino2tensorflow #OpenVINO to TFLite export